requirements.txt
opencv-python~=4.11.0.86
numpy~=2.2.3
Pillow~=9.2.0
mediapipe